+++ src/main/java/woowacourse/shoppingcart/controller/CartItemController.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.ui;
+package woowacourse.shoppingcart.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -7,7 +7,7 @@ import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import woowacourse.shoppingcart.domain.Cart;
 import woowacourse.shoppingcart.domain.Product;
 import woowacourse.shoppingcart.dto.Request;
-import woowacourse.shoppingcart.application.CartService;
+import woowacourse.shoppingcart.service.CartService;
 
 import java.net.URI;
 import java.util.List;
