+++ src/main/java/woowacourse/shoppingcart/controller/CustomerController.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.ui;
+package woowacourse.shoppingcart.controller;
 
 import org.springframework.web.bind.annotation.RestController;
