+++ src/main/java/woowacourse/shoppingcart/controller/OrderController.java
@@ -1,11 +1,11 @@
-package woowacourse.shoppingcart.ui;
+package woowacourse.shoppingcart.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import woowacourse.shoppingcart.dto.OrderRequest;
 import woowacourse.shoppingcart.domain.Orders;
-import woowacourse.shoppingcart.application.OrderService;
+import woowacourse.shoppingcart.service.OrderService;
 
 import javax.validation.Valid;
 import java.net.URI;
