+++ src/main/java/woowacourse/shoppingcart/controller/ProductController.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.ui;
+package woowacourse.shoppingcart.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import woowacourse.shoppingcart.domain.Product;
 import woowacourse.shoppingcart.dto.Request;
-import woowacourse.shoppingcart.application.ProductService;
+import woowacourse.shoppingcart.service.ProductService;
 
 import java.net.URI;
 import java.util.List;
